requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==4.2.2
altgraph==0.17.3
anyio==3.7.1
appnope==0.1.3
argcomplete==1.10.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
async-lru==2.0.3
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.8.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
chroma-hnswlib==0.7.1
chromadb==0.4.3
click==8.1.6
coloredlogs==15.0.1
comm==0.1.3
compressed-rtf==1.0.6
cryptography==41.0.2
dataclasses-json==0.5.13
DateTime==5.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docx2txt==0.8
ebcdic==1.1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
extract-msg==0.28.7
fastapi==0.99.1
fastjsonschema==2.18.0
filetype==1.2.0
flatbuffers==23.5.26
fqdn==1.5.1
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
h11==0.14.0
httplib2==0.22.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
IMAPClient==2.1.0
importlib-metadata==6.7.0
importlib-resources==6.0.0
ipykernel==6.23.3
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.23.0
langchain==0.0.247
langsmith==0.0.15
lxml==4.9.3
macholib==1.16.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
modulegraph==0.19.5
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.2
nbformat==5.9.1
nest-asyncio==1.5.6
nltk==3.8.1
notebook==7.0.0
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.24.2
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.15.1
openai==0.27.1
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parsedatetime==2.6
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20201018
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
posthog==3.0.1
prometheus-client==0.17.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
py2app==0.28.6
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.12
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pypandoc==1.11
pyparsing==3.1.0
PyPika==0.48.9
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
redis==4.5.1
referencing==0.30.0
regex==2023.6.3
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.9.2
rsa==4.9
selenium==4.10.0
semver==3.0.1
Send2Trash==1.8.2
six==1.12.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.8.1
SQLAlchemy==2.0.19
stack-data==0.6.2
starlette==0.27.0
streamlit==1.25.0
streamlit-chat==0.0.2.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.2.0
terminado==0.17.1
textract==1.6.5
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.3
typing-inspect==0.9.0
typing_extensions==4.7.0
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.8.7
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.1
uvloop==0.17.0
validators==0.20.0
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
widgetsnbextension==4.0.8
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
zope.interface==6.0
